add setting of qiniu enable

+++ src/controller/center.js
@@ -48,10 +48,19 @@ module.exports = class extends Base {
         let sk = this.config('site').qiniusk.value;
         let scope  = this.config('site').qiniuscope.value;
         let staticdomain = this.config('site').staticdomain.value;
+        let qiniuenable = this.config('site').qiniuenable.value == 0 ? true : false;//是否启用七牛云存储。
         let name = file.name;
-        let filePath = await qiniuCloud.saveFile(ak,sk,scope,file.path,'static_',path.extname(name));
-        let realPath = staticdomain+'/'+filePath;
-        fs.unlinkSync(file.path);//删除源文件。
+        let realPath = "";
+        if(qiniuenable){
+            let filePath = await qiniuCloud.saveFile(ak,sk,scope,file.path,'static_',path.extname(name));
+            realPath = staticdomain+'/'+filePath;
+            fs.unlinkSync(file.path);//删除源文件。
+        }else{
+            realPath = '/static/upload/user/' + name;
+            let filePath = path.join(think.ROOT_PATH, 'www', tempPath);
+            think.mkdir(path.dirname(filePath));
+            await rename(file.path, filePath);
+        }
         this.json({
             success: true,
             result: {
@@ -68,6 +77,8 @@ module.exports = class extends Base {
         let sk = this.config('site').qiniusk.value;
         let scope  = this.config('site').qiniuscope.value;
         let staticdomain = this.config('site').staticdomain.value;
+        let qiniuenable = this.config('site').qiniuenable.value == 0 ? true : false;//是否启用七牛云存储。
+
         let imgData = this.post('imgData');
         var base64Data = imgData.replace(/^data:image\/\w+;base64,/, "");
         var dataBuffer = new Buffer(base64Data, 'base64');
@@ -75,9 +86,14 @@ module.exports = class extends Base {
         let absPath = '/static/upload/user/' + fileName;
         let filePath = path.join(think.ROOT_PATH, 'www', absPath);
         fs.writeFileSync(filePath, dataBuffer);
-        let staticPath = await qiniuCloud.saveFile(ak,sk,scope,filePath,'static_');
-        fs.unlinkSync(filePath);//删除源文件。
-        let finalPath = staticdomain+'/'+staticPath;
+        let finalPath = "";
+        if(qiniuenable){
+            let staticPath = await qiniuCloud.saveFile(ak,sk,scope,filePath,'static_');
+            fs.unlinkSync(filePath);//删除源文件。
+            finalPath = staticdomain+'/'+staticPath;
+        }else{
+            finalPath = absPath;
+        }
         this.body = finalPath;
     }
     //====================站点属性设置--edi==========================t
